🐛 fix: 타입 불일치 수정

+++ src/main/java/no/gunbang/market/domain/auction/repository/AuctionRepositoryImpl.java
@@ -93,7 +93,7 @@ public class AuctionRepositoryImpl implements AuctionRepositoryCustom {
                 )
             );
         } else {
-            long maxBidderCount = Long.MAX_VALUE;
+            int maxBidderCount = Integer.MAX_VALUE;
             long maxAuctionId = Long.MAX_VALUE;
             builder.and(
                 Expressions.booleanTemplate(
